requirements.txt
bson==0.5.10
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
cors==1.0.1
filelock==3.7.1
Flask==2.1.3
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.2
future==0.18.2
gevent==21.12.0
greenlet==1.1.2
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.1
PyJWT==2.4.0
pymongo==4.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
requests==2.28.1
requests-file==1.5.1
six==1.16.0
tldextract==3.3.1
urllib3==1.26.10
Werkzeug==2.1.2
zipp==3.8.1
zope.event==4.5.0
zope.interface==5.4.0
